requirements.txt
appnope==0.1.2
asgiref==3.4.1
autopep8==1.5.7
backcall==0.2.0
boto3==1.19.2
botocore==1.22.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-cors-headers==3.8.0
django-rest-auth==0.9.5
django-storages==1.12.2
djangorestframework==3.12.4
entrypoints==0.3
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
matplotlib-inline==0.1.3
mysql==0.0.3
mysql-connector-python==8.0.23
mysqlclient==2.0.3
nest-asyncio==1.5.1
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.2.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
scikit-learn==1.0
scipy==1.7.1
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-103/six-1.15.0-py2.py3-none-any.whl
SQLAlchemy==1.4.26
sqlparse==0.4.2
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5